Pass the ID of the item that was clicked by adding it to an intent
When you use a category activity to display items in a list view, you’ll usually use the onItemClick() method to start another activity that displays details of the item the user clicked. To do this, you create an intent that starts the second activity. You then add the ID of the item that was clicked as extra information so that the second activity can use it when the activity starts.

Signed-off-by: bitlin <bitlin@outlook.com>